# apps/messaging-api/openapi-definition.yml
openapi: 3.0.3
info:
  title: OGCIO MessagingIE API
  description: API for OGCIO MessagingIE Service
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/messages/search:
    post:
      operationId: ListMessagesPost
      tags: *a1
      description: Returns all the messages for the requested organisation or the
        requested recipient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    - delivered
                isSeen:
                  type: string
                  enum:
                    - "true"
                    - "false"
                    - "0"
                    - "1"
                search:
                  type: string
                deleted:
                  type: string
                  enum:
                    - "true"
                    - "false"
                    - "0"
                    - "1"
                recipientUserId:
                  type: string
                  description: Either recipientUserId and organisationId are mandatory
                organisationId:
                  type: string
                  description: Either recipientUserId and organisationId are mandatory
                offset:
                  type: string
                  pattern: ^[0-9][0-9]*|undefined$
                  default: "0"
                  description: Indicates where to start fetching data or how many records to skip,
                    defining the initial position within the list
                limit:
                  type: string
                  default: "20"
                  pattern: ^([1-9]|100)|undefined$
                  description: Indicates the maximum number (100) of items that will be returned
                    in a single request
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      required:
                        - id
                        - subject
                        - createdAt
                        - threadName
                        - organisationId
                        - recipientUserId
                        - attachmentsCount
                      properties:
                        id:
                          type: string
                          description: Unique Id of the message
                        subject:
                          type: string
                          description: Subject
                        createdAt:
                          type: string
                          description: Creation date time
                        threadName:
                          anyOf:
                            - type: string
                              description: Thread Name used to group messages
                            - type: "null"
                        organisationId:
                          type: string
                          description: Organisation sender id
                        recipientUserId:
                          type: string
                          description: Unique id of the recipient
                        attachmentsCount:
                          type: integer
                          description: Number of attachments
                  metadata:
                    type: object
                    properties:
                      links:
                        type: object
                        required:
                          - self
                          - first
                          - last
                          - pages
                        properties:
                          self:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the request itself
                          next:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the next page of results in a paginated response.
                                  If there are no more results, this field may
                                  be omitted
                          prev:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the previous page of results in a paginated
                                  response. If there are no more results, this
                                  field may be omitted
                          first:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the first page of results in a paginated response
                          last:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the first page of results in a paginated response
                          pages:
                            type: object
                            description: It may contain a list of other useful URLs, e.g. one entry for
                              page:'page 1', 'page 2'
                            additionalProperties:
                              type: object
                              properties:
                                href:
                                  type: string
                        description: Object containing the links to the related endpoints
                      totalCount:
                        type: number
                        description: Number representing the total number of available items
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
                properties:
                  code:
                    type: string
                    description: Code used to categorize the error
                  detail:
                    type: string
                    description: Description of the error
                  requestId:
                    type: string
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                  name:
                    type: string
                    description: Name of the error type
                  validation:
                    type: array
                    items:
                      type: object
                      required:
                        - fieldName
                        - message
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                    description: List of the validation errors
                  validationContext:
                    type: string
                  statusCode:
                    type: number
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
                properties:
                  code:
                    type: string
                    description: Code used to categorize the error
                  detail:
                    type: string
                    description: Description of the error
                  requestId:
                    type: string
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                  name:
                    type: string
                    description: Name of the error type
                  validation:
                    type: array
                    items:
                      type: object
                      required:
                        - fieldName
                        - message
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                    description: List of the validation errors
                  validationContext:
                    type: string
                  statusCode:
                    type: number
...
  /api/v1/messages/{messageId}/events:
    get:
      tags: *a1
      description: Returns the list of events for the given message id
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: messageId
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      required:
                        - messageId
                        - eventType
                        - eventStatus
                        - data
                        - createdAt
                      properties:
                        messageId:
                          type: string
                          format: uuid
                          description: Message id
                        eventType:
                          type: string
                          description: Event type description
                        eventStatus:
                          type: string
                          description: Status for event type
                        data:
                          anyOf:
                            - type: object
                              required:
                                - messageId
                                - receiverFullName
                                - receiverPPSN
                                - receiverUserId
                                - subject
                                - language
                                - plainText
                                - transports
                                - scheduledAt
                                - organisationName
                                - security
                              properties:
                                messageId:
                                  type: string
                                  description: Unique id of the related message
                                receiverFullName:
                                  type: string
                                  description: Full name of the recipient
                                receiverPPSN:
                                  type: string
                                  description: PPSN of the recipient
                                receiverUserId:
                                  type: string
                                  description: User id of recipient
                                subject:
                                  type: string
                                  description: Subject of the related message
                                language:
                                  type: string
                                  description: Language of the related message
                                excerpt:
                                  type: string
                                  description: Excerpt of the related message
                                richText:
                                  type: string
                                  description: Rich text content of the related message
                                plainText:
                                  type: string
                                  description: Plain text context of the related message
                                threadName:
                                  type: string
                                  description: Thread name of the related message
                                transports:
                                  type: array
                                  items:
                                    type: string
                                  description: Selected transports to send the message
                                scheduledAt:
                                  type: string
                                  format: date-time
                                  description: Date and time which describes when the message has to be sent
                                senderUserId:
                                  type: string
                                  description: Unique user id of the sender
                                senderFullName:
                                  type: string
                                  description: Full name of the sender
                                senderPPSN:
                                  type: string
                                  description: PPSN of the sender
                                senderApplicationId:
                                  type: string
                                  description: Unique id of the M2M application that sent the message
                                organisationName:
                                  type: string
                                  description: Organisation related to the sender
                                security:
                                  type: string
                                  enum:
                                    - confidential
                                    - public
                                  description: Confidentiality level of the message
                            - type: object
                              required:
                                - messageId
                                - jobId
                              properties:
                                messageId:
                                  type: string
                                  description: Unique id of the related message
                                jobId:
                                  type: string
                                  description: Unique id of the job
                            - type: object
                              required:
                                - messageId
                              properties:
                                messageId:
                                  type: string
                                  description: Unique id of the related message
                        createdAt:
                          type: string
                          format: date-time
                          description: Date and time which describes when the event has been recorded
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
                properties:
                  code:
                    type: string
                    description: Code used to categorize the error
                  detail:
                    type: string
                    description: Description of the error
                  requestId:
                    type: string
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                  name:
                    type: string
                    description: Name of the error type
                  validation:
                    type: array
                    items:
                      type: object
                      required:
                        - fieldName
                        - message
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                    description: List of the validation errors
                  validationContext:
                    type: string
                  statusCode:
                    type: number
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
                properties:
                  code:
                    type: string
                    description: Code used to categorize the error
                  detail:
                    type: string
                    description: Description of the error
                  requestId:
                    type: string
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                  name:
                    type: string
                    description: Name of the error type
                  validation:
                    type: array
                    items:
                      type: object
                      required:
                        - fieldName
                        - message
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                    description: List of the validation errors
                  validationContext:
                    type: string
                  statusCode:
                    type: number
...
security:
  - bearerAuth: []